+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Add H264 Ios Stream Link to Camera Device
+
 ## 2.2.3 (2022-12-04)
 
 * (theimo1221) Add BlueIris settings and move settings interface to own files